+++ archive.md
@@ -1,0 +1,21 @@
+# Archive
+
+This doc is meant to capture some things that I've archived from my personal
+site in the past, I may bring these back!
+
+### Random section
+
+Collection of random things:
+
+```tsx
+<Heading is='h3' className='my-3'>
+  Random
+</Heading>
+<Text className='mb-2'>
+  I sometimes hack on random ideas that don't really need their own domain or might not really be worth throwing
+  in their own repo, you can find some of those in my Random part of the site.
+</Text>
+<Link href='/random'>Check them out</Link>
+```
+
+Pages are in `./old-app/(default)/random`
