--- paths.json
@@ -1,9 +1,0 @@
-[
-  "/notebook/2023/march/sandbox",
-  "/notebook/2020/june/the-art-of-note-taking",
-  "/notebook/2020/june/building-a-developer-toolbox",
-  "/notebook/2020/may/design-tokens-and-theme",
-  "/notebook/2020/may/the-state-of-monorepos",
-  "/notebook/2020/november/2020",
-  "/notebook"
-]
